Changes to npc-server: - level.issparringzone - Fixed an issue with npc.save[] sending to players - Tweaked the timer to be more accurate, and report better statistics.
Other changes:
- Changed around many files with the ScriptBinding's to try to separate it from being intertwined with the v8 implementation / gserver.
- Potentially fix segmentation fault on exit. V8 should now clean all constructors we bind too, should fix multiple-server support but it has not been tested.

+++ server/src/TPlayerNC.cpp
@@ -237,7 +237,7 @@ bool TPlayer::msgPLI_NC_NPCADD(CString& pPacket)
 		return true;
 	}
 
-	TNPC *newNpc = server->addServerNpc(strtoint(npcId), strtofloat(npcX), strtofloat(npcY), level, true);
+	TNPC *newNpc = server->addServerNpc(strtoint(npcId), (float)strtofloat(npcX), (float)strtofloat(npcY), level, true);
 
 	if (newNpc != nullptr)
 	{
